feat: class product and clean webscraper
