+++ webscraper/bin/webscraper.dart
@@ -1,5 +1,6 @@
 import 'package:webscraper/webscraper.dart' as webscraper;
+import 'package:webscraper/product.dart';
 
-Future<List<webscraper.Product>> main(List<String> arguments) {
+Future<List<Product>> main(List<String> arguments) {
   return webscraper.fetchAmazonProducts("steak", "monoprix");
 }
